# Added pre-processed required affinity terms to scheduler's PodInfo type.