**What type of PR is this?**

/kind feature

**What this PR does / why we need it**:
Adds processed required affinity terms to the scheduler's PodInfo. This brings 13% performance improvement for required anti-affinity benchmarks.

before
```
BenchmarkScheduling/5000Nodes/5000Pods-12                          1000  1411689   ns/op
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12           1000  8155520   ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/5000Pods-12               1000  12699040  ns/op
BenchmarkSchedulingPreferredPodAffinity/5000Nodes/5000Pods-12      1000  11805901  ns/op
BenchmarkSchedulingPreferredPodAntiAffinity/5000Nodes/5000Pods-12  1000  12808495  ns/op
BenchmarkSchedulingNodeAffinity/5000Nodes/5000Pods-12              1000  2725989   ns/op
```
after 
```
BenchmarkScheduling/5000Nodes/5000Pods-12                          1000  1419240   ns/op
BenchmarkSchedulingPodAntiAffinity/5000Nodes/1000Pods-12           1000  7225467   ns/op
BenchmarkSchedulingPodAffinity/5000Nodes/5000Pods-12               1000  12679068  ns/op
BenchmarkSchedulingPreferredPodAffinity/5000Nodes/5000Pods-12      1000  10958616  ns/op
BenchmarkSchedulingPreferredPodAntiAffinity/5000Nodes/5000Pods-12  1000  11353006  ns/op
BenchmarkSchedulingNodeAffinity/5000Nodes/5000Pods-12              1000  2785566   ns/op
```

**Which issue(s) this PR fixes**:
Part of #91058

**Special notes for your reviewer**:

**Does this PR introduce a user-facing change?**:

```release-note
NONE
```