@ahg-g 2020-05-14
> Some of the comments are for pre-existing code. Feel free to fix in this PR or followup PR.

Thanks Wei, fixed the function name and removed the check on NodeName being set in the update function. I amended since the changes are really minor.